requirements.txt
python-dotenv
bluepy
paho-mqtt
